+++ src/pages/Error.jsx
@@ -1,0 +1,8 @@
+
+function Error () {
+  return (
+    <div>Page not found</div>
+  )
+}
+
+export default Error
