+++ src/pages/Singin.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Singin = () => {
+  return (
+    <div>Singin</div>
+  )
+}
+
+export default Singin
