+++ src/pages/Singup.jsx
@@ -1,0 +1,8 @@
+
+function Singup () {
+  return (
+    <div>SingUp</div>
+  )
+}
+
+export default Singup
